# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Vowed API",
    "version": "0.1.0",
    "description": "Backend for the Vowed Android app. All endpoints except /v1/health, /v1/meta and /v1/auth/* need `Authorization: Bearer <token>` from /v1/auth/verify. u64 amounts are decimal strings. Transaction builders return unsigned base64 transactions: the app must decode and check them before asking the wallet to sign."
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "paths": {
...
    "/v1/auth/nonce": {
      "post": {
        "summary": "Get a sign-in nonce and the exact payload to pass to the wallet's sign-in",
        "tags": [
          "auth"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "wallet": {
                    "type": "string"
                  }
                },
                "required": [
                  "wallet"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "nonce": {
                      "type": "string"
                    },
                    "domain": {
                      "type": "string"
                    },
                    "statement": {
                      "type": "string"
                    },
                    "uri": {
                      "type": "string"
                    },
                    "issuedAt": {
                      "type": "string"
                    },
                    "expirationTime": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "nonce",
                    "domain",
                    "statement",
                    "uri",
                    "issuedAt",
                    "expirationTime"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/v1/auth/verify": {
      "post": {
        "summary": "Exchange the wallet's signed sign-in message for a short-lived token",
        "tags": [
          "auth"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "wallet": {
                    "type": "string"
                  },
                  "message": {
                    "type": "string",
                    "maxLength": 4096
                  },
                  "signature": {
                    "type": "string",
                    "maxLength": 200
                  }
                },
                "required": [
                  "wallet",
                  "message",
                  "signature"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "token": {
                      "type": "string"
                    },
                    "expiresAt": {
                      "type": "number"
                    },
                    "wallet": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "token",
                    "expiresAt",
                    "wallet"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
}